requirements.txt
streamlit
fastapi
uvicorn
pydantic
requests
ollama
langchain
langchain-community
twilio
python-dotenv
langgraph
langchain-groq
